+++ crates/stellar-registry-cli/CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.6](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-registry-cli-v0.0.5...stellar-registry-cli-v0.0.6) - 2025-07-08
+
+### Other
+
+- add publish documentation, remove unecessary config docs at top level ([#121](https://github.com/AhaLabs/scaffold-stellar/pull/121))
+
+## [0.0.5](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-registry-cli-v0.0.4...stellar-registry-cli-v0.0.5) - 2025-06-26
+
+### Fixed
+
+- *(registry-cli)* update mainnet instructions to inform about security practices and docs ([#112](https://github.com/AhaLabs/scaffold-stellar/pull/112))
+
+## [0.0.4](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-registry-cli-v0.0.3...stellar-registry-cli-v0.0.4) - 2025-06-18
+
+### Fixed
+
+- update readmes to point to current directory instead of project ([#109](https://github.com/AhaLabs/scaffold-stellar/pull/109))
+- [registry-cli] add wasm_name arg to publish  ([#90](https://github.com/AhaLabs/scaffold-stellar/pull/90))
+
 ## [0.0.3](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-registry-cli-v0.0.2...stellar-registry-cli-v0.0.3) - 2025-06-11
 
 ### Fixed