+++ crates/stellar-scaffold-cli/CHANGELOG.md
@@ -7,6 +7,50 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.5](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-scaffold-cli-v0.0.4...stellar-scaffold-cli-v0.0.5) - 2025-07-08
+
+### Added
+
+- use global args and printer in build / watch ([#124](https://github.com/AhaLabs/scaffold-stellar/pull/124))
+- add --no-prompt to upgrade ([#122](https://github.com/AhaLabs/scaffold-stellar/pull/122))
+
+### Fixed
+
+- init not copying fungible token to proper directory when relative path used ([#127](https://github.com/AhaLabs/scaffold-stellar/pull/127))
+- init flaky test failure ([#130](https://github.com/AhaLabs/scaffold-stellar/pull/130))
+- continue building clients on error ([#88](https://github.com/AhaLabs/scaffold-stellar/pull/88))
+
+### Other
+
+- Build clients race condition ([#126](https://github.com/AhaLabs/scaffold-stellar/pull/126))
+- add/update NFT token example on init ([#136](https://github.com/AhaLabs/scaffold-stellar/pull/136))
+
+## [0.0.4](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-scaffold-cli-v0.0.3...stellar-scaffold-cli-v0.0.4) - 2025-06-26
+
+### Added
+
+- init retrieves fresh FT interface ([#95](https://github.com/AhaLabs/scaffold-stellar/pull/95))
+
+### Fixed
+
+- *(registry-cli)* update mainnet instructions to inform about security practices and docs ([#112](https://github.com/AhaLabs/scaffold-stellar/pull/112))
+
+### Other
+
+- upgrade a project to a scaffold project ([#114](https://github.com/AhaLabs/scaffold-stellar/pull/114))
+
+## [0.0.3](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-scaffold-cli-v0.0.2...stellar-scaffold-cli-v0.0.3) - 2025-06-18
+
+### Added
+
+- add Cargo metadata ([#108](https://github.com/AhaLabs/scaffold-stellar/pull/108))
+
+### Fixed
+
+- update readmes to point to current directory instead of project ([#109](https://github.com/AhaLabs/scaffold-stellar/pull/109))
+- only copy contents to output location ([#107](https://github.com/AhaLabs/scaffold-stellar/pull/107))
+- ensure account gets funded even if identity already exists ([#106](https://github.com/AhaLabs/scaffold-stellar/pull/106))
+
 ## [0.0.2](https://github.com/AhaLabs/scaffold-stellar/compare/stellar-scaffold-cli-v0.0.1...stellar-scaffold-cli-v0.0.2) - 2025-06-11
 
 ### Added